// apps/backend/src/fiches/enums/fiche-statut.enum.ts
export enum FicheStatut {
  BROUILLON = 'BROUILLON',
  EN_COURS = 'EN_COURS',
  FINALISEE = 'FINALISEE',
  IMPRIMEE = 'IMPRIMEE',
  REMISE_CHAUFFEUR = 'REMISE_CHAUFFEUR',
  RECEPTION_CONFIRME = 'RECEPTION_CONFIRME',
  CHARGEMENT = 'CHARGEMENT',
  EN_CIRCULATION = 'EN_CIRCULATION',
  CLOTUREE = 'CLOTUREE',
  ANNULEE = 'ANNULEE',
  EN_ATTENTE = 'EN_ATTENTE',
  }